backend: refactoring SpdxService: moved license handling to LicenseHandler.java
