+++ occtet-nats/occtet-nats-spdx-service/src/main/java/eu/occtet/boc/spdx/service/SpdxService.java
@@ -36,11 +36,9 @@ import org.apache.logging.log4j.Logger;
 import org.spdx.core.InvalidSPDXAnalysisException;
 import org.spdx.core.TypedValue;
 import org.spdx.jacksonstore.MultiFormatStore;
-import org.spdx.library.LicenseInfoFactory;
 import org.spdx.library.SpdxModelFactory;
 import org.spdx.library.model.v2.*;
 import org.spdx.library.model.v2.license.*;
-import org.spdx.library.model.v3_0_1.expandedlicensing.ListedLicense;
 import org.spdx.storage.simple.InMemSpdxStore;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -71,11 +69,7 @@ public class SpdxService extends ProgressReportingService  {
     @Autowired
     private InventoryItemService inventoryItemService;
     @Autowired
-    private LicenseService licenseService;
-    @Autowired
     private FileService fileService;
-    @Autowired
-    private LicenseRepository licenseRepository;
     @Autowired
     private InventoryItemRepository inventoryItemRepository;
     @Autowired
@@ -88,6 +82,8 @@ public class SpdxService extends ProgressReportingService  {
     private CopyrightRepository copyrightRepository;
     @Autowired
     private CleanUpService cleanUpService;
+    @Autowired
+    private LicenseHandler licenseHandler;
 
     private Collection<ExtractedLicenseInfo> licenseInfosExtractedSpdxDoc = new ArrayList<>();
 
@@ -308,7 +304,7 @@ public class SpdxService extends ProgressReportingService  {
             spdxPkgLicense = spdxPackage.getLicenseDeclared();
         }
 
-        List<License> pkgLicenses = createLicenses(spdxPkgLicense, licenseCache);
+        List<License> pkgLicenses = licenseHandler.createLicenses(spdxPkgLicense, licenseCache, licenseInfosExtractedSpdxDoc);
         if(component.getLicenses() != null) {
             //make double sure there are no doubles
             Set<License> lSet= new HashSet<>( component.getLicenses());
@@ -381,76 +377,6 @@ public class SpdxService extends ProgressReportingService  {
         }
 
         return inventoryItem;
-    }
-
-    private List<License> createLicenses(AnyLicenseInfo spdxLicenseInfo, Map<String, License> licenseCache)
-            throws InvalidSPDXAnalysisException {
-
-        Set<License> allLicenses = new HashSet<>();
-        List<AnyLicenseInfo> allLicenseInfo = new ArrayList<>();
-        parseLicenseText(spdxLicenseInfo, allLicenseInfo);
-
-        for (AnyLicenseInfo individualLicenseInfo : allLicenseInfo) {
-            String licenseId = "";
-            String licenseText = "";
-            boolean isListed = false;
-
-            if (individualLicenseInfo instanceof SpdxListedLicense) {
-                ListedLicense license = LicenseInfoFactory.getListedLicenseById(individualLicenseInfo.getId());
-                licenseId = license.getId();
-                licenseText = license.getLicenseText();
-                isListed = true;
-            } else if (individualLicenseInfo instanceof ExtractedLicenseInfo) {
-                Optional<ExtractedLicenseInfo> extracted = licenseInfosExtractedSpdxDoc.stream()
-                        .filter(s -> s.getLicenseId().equals(individualLicenseInfo.getId())).findFirst();
-                if (extracted.isPresent()) {
-                    licenseId = extracted.get().getId();
-                    licenseText = extracted.get().getExtractedText();
-                }
-            }
-
-            if (licenseId.isEmpty()) licenseId = "Unknown";
-
-            License licenseEntity = licenseCache.get(licenseId);
-
-            if (licenseEntity == null) {
-                licenseEntity = licenseService.findOrCreateLicense(licenseId, licenseText, licenseId);
-
-                if (isListed) {
-                    licenseEntity.setSpdx(true);
-                    licenseRepository.save(licenseEntity);
-                }
-
-                // Cache Put
-                licenseCache.put(licenseId, licenseEntity);
-            }
-
-            allLicenses.add(licenseEntity);
-        }
-
-        return allLicenses.stream().toList();
-    }
-
-    private void parseLicenseText(AnyLicenseInfo licenseInfo, List<AnyLicenseInfo> allLicenseInfos) throws InvalidSPDXAnalysisException {
-        switch (licenseInfo) {
-            case ConjunctiveLicenseSet conjunctiveLicenseSet -> {
-                for (AnyLicenseInfo member : conjunctiveLicenseSet.getMembers()) {
-                    parseLicenseText(member, allLicenseInfos);
-                }
-            }
-            case DisjunctiveLicenseSet disjunctiveLicenseSet -> {
-                for (AnyLicenseInfo member : disjunctiveLicenseSet.getMembers()) {
-                    parseLicenseText(member, allLicenseInfos);
-                }
-            }
-            case WithExceptionOperator withExceptionOperator -> parseLicenseText(withExceptionOperator.getLicense(), allLicenseInfos);
-            case SpdxListedLicense listed -> allLicenseInfos.add(listed);
-            case ExtractedLicenseInfo extracted -> allLicenseInfos.add(extracted);
-            //No action needed if there is no license
-            case SpdxNoneLicense ignored -> {
-            }
-            case null, default -> log.info("Encountered unknown license type: {}", licenseInfo);
-        }
     }
 
     private List<Copyright> parseFiles(SpdxPackage spdxPackage, InventoryItem inventoryItem) throws InvalidSPDXAnalysisException {
@@ -579,7 +505,7 @@ public class SpdxService extends ProgressReportingService  {
 
         AnyLicenseInfo concluded = snippet.getLicenseConcluded();
         if (concluded != null && !concluded.isNoAssertion(concluded) && !concluded.isNoAssertion(concluded)) {
-            List<License> licenses = createLicenses(concluded, licenseCache);
+            List<License> licenses = licenseHandler.createLicenses(concluded, licenseCache, licenseInfosExtractedSpdxDoc);
             for (License license : licenses) {
                 if (!component.getLicenses().contains(license)) {
                     component.addLicense(license);
@@ -672,7 +598,7 @@ public class SpdxService extends ProgressReportingService  {
             }
 
             if (fileLicense != null) {
-                List<License> licenses = createLicenses(fileLicense, licenseCache);
+                List<License> licenses = licenseHandler.createLicenses(fileLicense, licenseCache, licenseInfosExtractedSpdxDoc);
 
                 if (component.getLicenses() == null) {
                     component.setLicenses(new ArrayList<>());
